[CLEANUP] Declassify handle short events (#3669)
* new murder functions and structure

* fix bad refactor

* implementing new func for reveals

* adjusting how profiles display history for murders

* history displaying successfully

* some cleanup and adjusting get_murder_status_text

* adjust short event murder handling

* formatting

* fixing various bugs

* move all handling into short event object

* event related condition lists moved to json

* moving funcs into future event file

* move short event finding into generate_events

* some refactoring

* disentangling to avoid circular imports

oh god oh lord
short event generation now has it's own file away from the group
future event prep and trigger functions have their own singleton now

* circular imports still killing me

* separate trigger from future event class

* changing these to their enums

* fixing assorted bugs

* typing and some refactoring for readability

* updating and adding new tests

* tests for cat inclusion

* fixing merge changes

* refactoring for readability

* fixing up the future event trigger

* formatting

* specifying exceptions

* reorganising and adding documentation

* add back some documentation

* remove old arg

* remove accidental class method creation?

* fix mistyped func

* fix misassigned func

* merging in stuff from moved files

* moving into constants

* remove mismerged lines

* remove unused import

* fix mismerge

* fixing lost merged changes

* fixing hidden murder merge

* fixing merge issues

* suggested changes

* carrying over merge changes

* fix merge issues

* fix scar lists

* Update short_event_generation.py

* formatter

* fixing self.new_cats

* fix captilization

* add event repeat prevention

* merge fix

* fix collar attribute

* Update test_handle_short_events.py

* merge in CAMP changes

* formatter

* fix mismerge

* timeskip test change
