+++ scripts/clan_resources/herb/herb.py
@@ -3,8 +3,6 @@ import os
 import ujson
 import i18n
 
-from scripts.game_structure.localization import load_lang_resource
-
 
 class Herb:
     def __init__(self, herb_name):
