+++ scripts/game_structure/constants.py
@@ -1,6 +1,11 @@
+import tomllib
+
 from pygame import Cursor, image, SYSTEM_CURSOR_ARROW
 import ujson
-import tomllib
+
+# these scripts don't import any clangen scripts into themselves, so it's okay for them to be imported here
+from scripts.clan_resources.herb.herb import HERBS
+from scripts.clan_resources.supply import Supply
 
 from scripts.screens.enums import GameScreen
 
@@ -116,6 +121,19 @@ EVENT_ALLOWED_CONDITIONS = [
     "heat exhaustion",
     "stomachache",
     "constant nightmares",
+]
+
+SUPPLY_TYPES = ["fresh_kill", "all_herb", "any_herb"]
+SUPPLY_TYPES.extend(HERBS)
+
+SUPPLY_TRIGGERS = ["always", *Supply]
+
+SUPPLY_ADJUSTMENTS = [
+    "reduce_eighth",
+    "reduce_quarter",
+    "reduce_half",
+    "reduce_full",
+    "increase_#",
 ]
 
 with open("resources/game_config.toml", "r", encoding="utf-8") as read_file:
